PlayerValues refactoring
PlayerValues is a mapping from player to "value". Value is not the same as score: it only reflects how the player finished compared to the other players.

PlayerValues instances can be generated from scores or from tier lists.

MCTS has a separate NodeValues class that implements PlayerValues and adds functionality for agreggating values from multiple sources.