Modify MCTS action score definition
Alphazero uses zero for the predicted value component (q) of unvisited nodes.

This seems to give an excessive boost to visited nodes over unvisited nodes. Consider nodes A and B with very similar priors. A is visited first. Now A will be visited again some number of times before B is visited for the first time, since A's predicted value is non-zero while B's is still zero, even if once both are visited, their predicted values will be close enough that they are visited the same number of times after that point.

It seems better to use an estimated value for unvisited nodes instead of zero. This change computes that estimated value as the parent node's predicted value times the ratio of the unvisited action's prior logit to the max logit among all peer actions.